Src/AjBase/Sql/Parser.cs
namespace AjBase.Sql
{
    using System;
    using System.Collections;
    using System.Collections.Generic;
    using System.IO;
    using System.Linq;
    using System.Text;

    public class Parser
    {
        private Lexer lexer;

        public Parser(string text)
            : this(new Lexer(text))
        {
        }

        public Parser(TextReader reader)
            : this(new Lexer(reader))
        {
        }

        public Parser(Lexer lexer)
        {
            this.lexer = lexer;
        }

        public ICommand ParseCommand()
        {
            Token token = lexer.NextToken();

            if (token == null)
                return null;

            if (IsToken(token, "create", TokenType.Name))
                return ParseCreate();

            throw new ParserException(string.Format("Unexpected '{0}'", token.Value));
        }

        private ICommand ParseCreate()
        {
            Token token = this.lexer.NextToken();

            if (token == null)
                throw new ParserException(string.Format("Unexpected end of input"));

            if (IsName(token, "database"))
                return ParseCreateDatabase();

            throw new ParserException(string.Format("Unexpected '{0}'", token.Value));
        }

        private ICommand ParseCreateDatabase()
        {
            Token token = this.lexer.NextToken();

            if (token == null)
                throw new ParserException(string.Format("Unexpected end of input"));

            if (token.TokenType == TokenType.Name)
                return new CreateDatabaseCommand(token.Value);

            throw new ParserException(string.Format("Unexpected '{0}'", token.Value));
        }

        private static bool IsName(Token token, string value)
        {
            return IsToken(token, value, TokenType.Name);
        }

        private static bool IsToken(Token token, string value, TokenType type)
        {
            if (token == null)
                return false;

            if (token.TokenType != type)
                return false;

            if (type == TokenType.Name)
                return token.Value.Equals(value, StringComparison.InvariantCultureIgnoreCase);

            return token.Value.Equals(value);
        }
    }
}
